+++ tests/js/route.test.js
@@ -9,8 +9,6 @@ const defaultWindow = {
 
 const defaultZiggy = {
     baseUrl: 'https://ziggy.dev',
-    baseProtocol: 'https',
-    baseDomain: 'ziggy.dev',
     basePort: null,
     defaultParameters: { locale: 'en' },
     namedRoutes: {
@@ -345,7 +343,6 @@ describe('route()', () => {
 
     test('can generate a URL with a port', () => {
         global.Ziggy.baseUrl = 'https://ziggy.dev:81';
-        global.Ziggy.baseDomain = 'ziggy.dev';
         global.Ziggy.basePort = 81;
 
         // route with no parameters
@@ -393,8 +390,6 @@ describe('route()', () => {
     test('can accept a custom Ziggy configuration object', () => {
         const config = {
             baseUrl: 'http://notYourAverage.dev',
-            baseProtocol: 'http',
-            baseDomain: 'notYourAverage.dev',
             basePort: false,
             defaultParameters: { locale: 'en' },
             namedRoutes: {
@@ -512,8 +507,6 @@ describe('current()', () => {
 
         const config = {
             baseUrl: 'https://ziggy.dev',
-            baseProtocol: 'https',
-            baseDomain: 'ziggy.dev',
             basePort: false,
             namedRoutes: {
                 'events.index': {